+++ providers/palette-context.tsx
@@ -15,6 +15,7 @@ type PaletteContextType = {
   palette: string,
   setPalette: Dispatch<SetStateAction<string>>
 }
+
 export const PaletteContext = createContext<PaletteContextType | undefined>(undefined);
 
 const PaletteContextProvider = ({ children }: PaletteContextProviderProps) => {
@@ -28,14 +29,6 @@ const PaletteContextProvider = ({ children }: PaletteContextProviderProps) => {
       {children}
     </PaletteContext.Provider>
   )
-}
-
-export const usePaletteContext = () => {
-  const context = useContext(PaletteContext)
-  if (context === undefined) {
-    throw new Error(`usePalette must be used within a PaletteContextProvider`)
-  }
-  return context
 }
 
 export default PaletteContextProvider;
